requirements.txt
pytest
requests
litellm
openai
python-dotenv
pytest-asyncio
httpx
httpx[http2]